+++ packages/react-router/__tests__/navigate-test.tsx
@@ -31,6 +31,7 @@ describe("<Navigate>", () => {
         );
       });
 
+      // @ts-expect-error
       expect(renderer.toJSON()).toMatchInlineSnapshot(`
         <h1>
           About
@@ -53,6 +54,7 @@ describe("<Navigate>", () => {
         );
       });
 
+      // @ts-expect-error
       expect(renderer.toJSON()).toMatchInlineSnapshot(`
         <h1>
           About
@@ -75,6 +77,7 @@ describe("<Navigate>", () => {
         );
       });
 
+      // @ts-expect-error
       expect(renderer.toJSON()).toMatchInlineSnapshot(`
         <h1>
           About
@@ -97,6 +100,7 @@ describe("<Navigate>", () => {
         );
       });
 
+      // @ts-expect-error
       expect(renderer.toJSON()).toMatchInlineSnapshot(`
         <h1>
           About
@@ -125,6 +129,7 @@ describe("<Navigate>", () => {
         );
       });
 
+      // @ts-expect-error
       expect(renderer.toJSON()).toMatchInlineSnapshot(`
         <h1>
           About
@@ -156,6 +161,7 @@ describe("<Navigate>", () => {
         );
       });
 
+      // @ts-expect-error
       expect(renderer.toJSON()).toMatchInlineSnapshot(`
         <h1>
           About
@@ -200,6 +206,7 @@ describe("<Navigate>", () => {
         );
       });
 
+      // @ts-expect-error
       expect(renderer.toJSON()).toMatchInlineSnapshot(`
         <h1>
           Home
@@ -225,6 +232,7 @@ describe("<Navigate>", () => {
         );
       });
 
+      // @ts-expect-error
       expect(renderer.toJSON()).toMatchInlineSnapshot(`
         <h1>
           Destination
@@ -250,6 +258,7 @@ describe("<Navigate>", () => {
         );
       });
 
+      // @ts-expect-error
       expect(renderer.toJSON()).toMatchInlineSnapshot(`
         <h1>
           Contacts
@@ -272,6 +281,7 @@ describe("<Navigate>", () => {
         );
       });
 
+      // @ts-expect-error
       expect(renderer.toJSON()).toMatchInlineSnapshot(`
         <h1>
           Contacts
@@ -297,6 +307,7 @@ describe("<Navigate>", () => {
         );
       });
 
+      // @ts-expect-error
       expect(renderer.toJSON()).toMatchInlineSnapshot(`
         <h1>
           Contacts
@@ -328,6 +339,7 @@ describe("<Navigate>", () => {
         );
       });
 
+      // @ts-expect-error
       expect(renderer.toJSON()).toMatchInlineSnapshot(`
         <h1>
           Contacts
@@ -359,6 +371,7 @@ describe("<Navigate>", () => {
         );
       });
 
+      // @ts-expect-error
       expect(renderer.toJSON()).toMatchInlineSnapshot(`
         <h1>
           Contacts
@@ -387,6 +400,7 @@ describe("<Navigate>", () => {
         );
       });
 
+      // @ts-expect-error
       expect(renderer.toJSON()).toMatchInlineSnapshot(`
         <h1>
           Destination
@@ -423,6 +437,7 @@ describe("<Navigate>", () => {
         );
       }
 
+      // @ts-expect-error
       expect(renderer.toJSON()).toMatchInlineSnapshot(`
         [
           <h1>
@@ -472,1114 +487,6 @@ describe("<Navigate>", () => {
           About
         </h1>
       </div>"
-    `);
-  });
-});
-
-function UseNavigateWrapper({
-  to,
-  relative,
-}: {
-  to: To;
-  relative: RelativeRoutingType;
-}) {
-  let navigate = useNavigate();
-  React.useEffect(() => {
-    navigate(to, { relative });
-  });
-  return null;
-}
-
-describe("useNavigate", () => {
-  describe("with an absolute href", () => {
-    it("navigates to the correct URL", () => {
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(
-          <MemoryRouter initialEntries={["/home"]}>
-            <Routes>
-              <Route path="home" element={<UseNavigateWrapper to="/about" />} />
-              <Route path="about" element={<h1>About</h1>} />
-            </Routes>
-          </MemoryRouter>
-        );
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          About
-        </h1>
-      `);
-    });
-  });
-
-  describe("with a relative href (relative=route)", () => {
-    it("navigates to the correct URL", () => {
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(
-          <MemoryRouter initialEntries={["/home"]}>
-            <Routes>
-              <Route
-                path="home"
-                element={<UseNavigateWrapper to="../about" />}
-              />
-              <Route path="about" element={<h1>About</h1>} />
-            </Routes>
-          </MemoryRouter>
-        );
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          About
-        </h1>
-      `);
-    });
-
-    it("handles upward navigation from an index routes", () => {
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(
-          <MemoryRouter initialEntries={["/home"]}>
-            <Routes>
-              <Route path="home">
-                <Route index element={<UseNavigateWrapper to="../about" />} />
-              </Route>
-              <Route path="about" element={<h1>About</h1>} />
-            </Routes>
-          </MemoryRouter>
-        );
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          About
-        </h1>
-      `);
-    });
-
-    it("handles upward navigation from inside a pathless layout route", () => {
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(
-          <MemoryRouter initialEntries={["/home"]}>
-            <Routes>
-              <Route element={<Outlet />}>
-                <Route
-                  path="home"
-                  element={<UseNavigateWrapper to="../about" />}
-                />
-              </Route>
-              <Route path="about" element={<h1>About</h1>} />
-            </Routes>
-          </MemoryRouter>
-        );
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          About
-        </h1>
-      `);
-    });
-
-    it("handles upward navigation from inside multiple pathless layout routes + index route", () => {
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(
-          <MemoryRouter initialEntries={["/home"]}>
-            <Routes>
-              <Route path="home">
-                <Route element={<Outlet />}>
-                  <Route element={<Outlet />}>
-                    <Route element={<Outlet />}>
-                      <Route
-                        index
-                        element={<UseNavigateWrapper to="../about" />}
-                      />
-                    </Route>
-                  </Route>
-                </Route>
-              </Route>
-              <Route path="about" element={<h1>About</h1>} />
-            </Routes>
-          </MemoryRouter>
-        );
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          About
-        </h1>
-      `);
-    });
-
-    it("handles upward navigation from inside multiple pathless layout routes + path route", () => {
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(
-          <MemoryRouter initialEntries={["/home/page"]}>
-            <Routes>
-              <Route path="home" element={<Outlet />}>
-                <Route element={<Outlet />}>
-                  <Route element={<Outlet />}>
-                    <Route element={<Outlet />}>
-                      <Route
-                        path="page"
-                        element={<UseNavigateWrapper to="../../about" />}
-                      />
-                    </Route>
-                  </Route>
-                </Route>
-              </Route>
-              <Route path="about" element={<h1>About</h1>} />
-            </Routes>
-          </MemoryRouter>
-        );
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          About
-        </h1>
-      `);
-    });
-
-    it("handles parent navigation from inside multiple pathless layout routes", () => {
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(
-          <MemoryRouter initialEntries={["/home/page"]}>
-            <Routes>
-              <Route
-                path="home"
-                element={
-                  <>
-                    <h1>Home</h1>
-                    <Outlet />
-                  </>
-                }
-              >
-                <Route element={<Outlet />}>
-                  <Route element={<Outlet />}>
-                    <Route element={<Outlet />}>
-                      <Route
-                        path="page"
-                        element={
-                          <>
-                            <h2>Page</h2>
-                            <UseNavigateWrapper to=".." />
-                          </>
-                        }
-                      />
-                    </Route>
-                  </Route>
-                </Route>
-              </Route>
-              <Route path="about" element={<h1>About</h1>} />
-            </Routes>
-          </MemoryRouter>
-        );
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          Home
-        </h1>
-      `);
-    });
-
-    it("handles relative navigation from nested index route", () => {
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(
-          <MemoryRouter initialEntries={["/layout/thing"]}>
-            <Routes>
-              <Route path="layout">
-                <Route path=":param">
-                  {/* redirect /layout/:param/ index routes to /layout/:param/dest */}
-                  <Route index element={<UseNavigateWrapper to="dest" />} />
-                  <Route path="dest" element={<h1>Destination</h1>} />
-                </Route>
-              </Route>
-            </Routes>
-          </MemoryRouter>
-        );
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          Destination
-        </h1>
-      `);
-    });
-  });
-
-  describe("with a relative href (relative=path)", () => {
-    it("navigates to the correct URL", () => {
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(
-          <MemoryRouter initialEntries={["/contacts/1"]}>
-            <Routes>
-              <Route path="contacts" element={<h1>Contacts</h1>} />
-              <Route
-                path="contacts/:id"
-                element={<UseNavigateWrapper to=".." relative="path" />}
-              />
-            </Routes>
-          </MemoryRouter>
-        );
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          Contacts
-        </h1>
-      `);
-    });
-
-    it("handles upward navigation from an index routes", () => {
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(
-          <MemoryRouter initialEntries={["/contacts/1"]}>
-            <Routes>
-              <Route path="contacts" element={<h1>Contacts</h1>} />
-              <Route path="contacts/:id">
-                <Route
-                  index
-                  element={<UseNavigateWrapper to=".." relative="path" />}
-                />
-              </Route>
-            </Routes>
-          </MemoryRouter>
-        );
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          Contacts
-        </h1>
-      `);
-    });
-
-    it("handles upward navigation from inside a pathless layout route", () => {
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(
-          <MemoryRouter initialEntries={["/contacts/1"]}>
-            <Routes>
-              <Route path="contacts" element={<h1>Contacts</h1>} />
-              <Route element={<Outlet />}>
-                <Route
-                  path="contacts/:id"
-                  element={<UseNavigateWrapper to=".." relative="path" />}
-                />
-              </Route>
-            </Routes>
-          </MemoryRouter>
-        );
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          Contacts
-        </h1>
-      `);
-    });
-
-    it("handles upward navigation from inside multiple pathless layout routes + index route", () => {
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(
-          <MemoryRouter initialEntries={["/contacts/1"]}>
-            <Routes>
-              <Route path="contacts" element={<h1>Contacts</h1>} />
-              <Route path="contacts/:id">
-                <Route element={<Outlet />}>
-                  <Route element={<Outlet />}>
-                    <Route element={<Outlet />}>
-                      <Route
-                        index
-                        element={<UseNavigateWrapper to=".." relative="path" />}
-                      />
-                    </Route>
-                  </Route>
-                </Route>
-              </Route>
-            </Routes>
-          </MemoryRouter>
-        );
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          Contacts
-        </h1>
-      `);
-    });
-
-    it("handles upward navigation from inside multiple pathless layout routes + path route", () => {
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(
-          <MemoryRouter initialEntries={["/contacts/1"]}>
-            <Routes>
-              <Route path="contacts" element={<Outlet />}>
-                <Route index element={<h1>Contacts</h1>} />
-                <Route element={<Outlet />}>
-                  <Route element={<Outlet />}>
-                    <Route element={<Outlet />}>
-                      <Route
-                        path=":id"
-                        element={<UseNavigateWrapper to=".." relative="path" />}
-                      />
-                    </Route>
-                  </Route>
-                </Route>
-              </Route>
-            </Routes>
-          </MemoryRouter>
-        );
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          Contacts
-        </h1>
-      `);
-    });
-
-    it("handles relative navigation from nested index route", () => {
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(
-          <MemoryRouter initialEntries={["/layout/thing"]}>
-            <Routes>
-              <Route path="layout">
-                <Route path=":param">
-                  {/* redirect /layout/:param/ index routes to /layout/:param/dest */}
-                  <Route
-                    index
-                    element={<UseNavigateWrapper to="dest" relative="path" />}
-                  />
-                  <Route path="dest" element={<h1>Destination</h1>} />
-                </Route>
-              </Route>
-            </Routes>
-          </MemoryRouter>
-        );
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          Destination
-        </h1>
-      `);
-    });
-
-    it("preserves search params and hash", () => {
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(
-          <MemoryRouter initialEntries={["/contacts/1"]}>
-            <Routes>
-              <Route path="contacts" element={<Contacts />} />
-              <Route
-                path="contacts/:id"
-                element={
-                  <UseNavigateWrapper to="..?foo=bar#hash" relative="path" />
-                }
-              />
-            </Routes>
-          </MemoryRouter>
-        );
-      });
-
-      function Contacts() {
-        let { search, hash } = useLocation();
-        return (
-          <>
-            <h1>Contacts</h1>
-            <p>
-              {search}
-              {hash}
-            </p>
-          </>
-        );
-      }
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        [
-          <h1>
-            Contacts
-          </h1>,
-          <p>
-            ?foo=bar
-            #hash
-          </p>,
-        ]
-      `);
-    });
-  });
-
-  it("is not stable across location changes", () => {
-    let router = createMemoryRouter(
-      [
-        {
-          path: "/",
-          Component: () => (
-            <>
-              <NavBar />
-              <Outlet />
-            </>
-          ),
-          children: [
-            {
-              path: "home",
-              element: <h1>Home</h1>,
-            },
-            {
-              path: "about",
-              element: <h1>About</h1>,
-            },
-          ],
-        },
-      ],
-      { initialEntries: ["/home"] }
-    );
-    let renderer: TestRenderer.ReactTestRenderer;
-    TestRenderer.act(() => {
-      renderer = TestRenderer.create(<RouterProvider router={router} />);
-    });
-
-    function NavBar() {
-      let count = React.useRef(0);
-      let navigate = useNavigate();
-      console.log("rendering NavBar");
-      React.useEffect(() => {
-        count.current++;
-      }, [navigate]);
-      return <p>{count.current}</p>;
-    }
-
-    expect(renderer.toJSON()).toMatchInlineSnapshot(`
-      [
-        <p>
-          0
-        </p>,
-        <h1>
-          Home
-        </h1>,
-      ]
-    `);
-
-    TestRenderer.act(() => {
-      router.navigate("/about");
-    });
-
-    expect(renderer.toJSON()).toMatchInlineSnapshot(`
-      [
-        <p>
-          1
-        </p>,
-        <h1>
-          About
-        </h1>,
-      ]
-    `);
-
-    TestRenderer.act(() => {
-      router.navigate("/home");
-    });
-
-    expect(renderer.toJSON()).toMatchInlineSnapshot(`
-      [
-        <p>
-          2
-        </p>,
-        <h1>
-          Home
-        </h1>,
-      ]
-    `);
-  });
-});
-
-function UseRouterNavigateWrapper({
-  to,
-  relative,
-}: {
-  to: To;
-  relative?: RelativeRoutingType;
-}) {
-  let navigate = useRouterNavigate();
-  React.useEffect(() => {
-    navigate(to, { relative });
-  });
-  return null;
-}
-
-describe("useRouterNavigate", () => {
-  describe("with an absolute href", () => {
-    it("navigates to the correct URL", () => {
-      let router = createMemoryRouter(
-        createRoutesFromElements(
-          <>
-            <Route
-              path="home"
-              element={<UseRouterNavigateWrapper to="/about" />}
-            />
-            <Route path="about" element={<h1>About</h1>} />
-          </>
-        ),
-        { initialEntries: ["/home"] }
-      );
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(<RouterProvider router={router} />);
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          About
-        </h1>
-      `);
-    });
-  });
-
-  describe("with a relative href (relative=route)", () => {
-    it("navigates to the correct URL", () => {
-      let router = createMemoryRouter(
-        createRoutesFromElements(
-          <>
-            <Route
-              path="home"
-              element={<UseRouterNavigateWrapper to="../about" />}
-            />
-            <Route path="about" element={<h1>About</h1>} />
-          </>
-        ),
-        { initialEntries: ["/home"] }
-      );
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(<RouterProvider router={router} />);
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          About
-        </h1>
-      `);
-    });
-
-    it("handles upward navigation from an index routes", () => {
-      let router = createMemoryRouter(
-        createRoutesFromElements(
-          <>
-            <Route path="home">
-              <Route
-                index
-                element={<UseRouterNavigateWrapper to="../about" />}
-              />
-            </Route>
-            <Route path="about" element={<h1>About</h1>} />
-          </>
-        ),
-        { initialEntries: ["/home"] }
-      );
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(<RouterProvider router={router} />);
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          About
-        </h1>
-      `);
-    });
-
-    it("handles upward navigation from inside a pathless layout route", () => {
-      let router = createMemoryRouter(
-        createRoutesFromElements(
-          <>
-            <Route element={<Outlet />}>
-              <Route
-                path="home"
-                element={<UseRouterNavigateWrapper to="../about" />}
-              />
-            </Route>
-            <Route path="about" element={<h1>About</h1>} />
-          </>
-        ),
-        { initialEntries: ["/home"] }
-      );
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(<RouterProvider router={router} />);
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          About
-        </h1>
-      `);
-    });
-
-    it("handles upward navigation from inside multiple pathless layout routes + index route", () => {
-      let router = createMemoryRouter(
-        createRoutesFromElements(
-          <>
-            <Route path="home">
-              <Route element={<Outlet />}>
-                <Route element={<Outlet />}>
-                  <Route element={<Outlet />}>
-                    <Route
-                      index
-                      element={<UseRouterNavigateWrapper to="../about" />}
-                    />
-                  </Route>
-                </Route>
-              </Route>
-            </Route>
-            <Route path="about" element={<h1>About</h1>} />
-          </>
-        ),
-        { initialEntries: ["/home"] }
-      );
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(<RouterProvider router={router} />);
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          About
-        </h1>
-      `);
-    });
-
-    it("handles upward navigation from inside multiple pathless layout routes + path route", () => {
-      let router = createMemoryRouter(
-        createRoutesFromElements(
-          <>
-            <Route path="home" element={<Outlet />}>
-              <Route element={<Outlet />}>
-                <Route element={<Outlet />}>
-                  <Route element={<Outlet />}>
-                    <Route
-                      path="page"
-                      element={<UseRouterNavigateWrapper to="../../about" />}
-                    />
-                  </Route>
-                </Route>
-              </Route>
-            </Route>
-            <Route path="about" element={<h1>About</h1>} />
-          </>
-        ),
-        { initialEntries: ["/home/page"] }
-      );
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(<RouterProvider router={router} />);
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          About
-        </h1>
-      `);
-    });
-
-    it("handles parent navigation from inside multiple pathless layout routes", () => {
-      let router = createMemoryRouter(
-        createRoutesFromElements(
-          <>
-            <Route
-              path="home"
-              element={
-                <>
-                  <h1>Home</h1>
-                  <Outlet />
-                </>
-              }
-            >
-              <Route element={<Outlet />}>
-                <Route element={<Outlet />}>
-                  <Route element={<Outlet />}>
-                    <Route
-                      path="page"
-                      element={
-                        <>
-                          <h2>Page</h2>
-                          <UseRouterNavigateWrapper to=".." />
-                        </>
-                      }
-                    />
-                  </Route>
-                </Route>
-              </Route>
-            </Route>
-            <Route path="about" element={<h1>About</h1>} />
-          </>
-        ),
-        { initialEntries: ["/home/page"] }
-      );
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(<RouterProvider router={router} />);
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          Home
-        </h1>
-      `);
-    });
-
-    it("handles relative navigation from nested index route", () => {
-      let router = createMemoryRouter(
-        createRoutesFromElements(
-          <>
-            <Route path="layout">
-              <Route path=":param">
-                {/* redirect /layout/:param/ index routes to /layout/:param/dest */}
-                <Route index element={<UseRouterNavigateWrapper to="dest" />} />
-                <Route path="dest" element={<h1>Destination</h1>} />
-              </Route>
-            </Route>
-          </>
-        ),
-        { initialEntries: ["/layout/thing"] }
-      );
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(<RouterProvider router={router} />);
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          Destination
-        </h1>
-      `);
-    });
-  });
-
-  describe("with a relative href (relative=path)", () => {
-    it("navigates to the correct URL", () => {
-      let router = createMemoryRouter(
-        createRoutesFromElements(
-          <>
-            <Route path="contacts" element={<h1>Contacts</h1>} />
-            <Route
-              path="contacts/:id"
-              element={<UseRouterNavigateWrapper to=".." relative="path" />}
-            />
-          </>
-        ),
-        { initialEntries: ["/contacts/1"] }
-      );
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(<RouterProvider router={router} />);
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          Contacts
-        </h1>
-      `);
-    });
-
-    it("handles upward navigation from an index routes", () => {
-      let router = createMemoryRouter(
-        createRoutesFromElements(
-          <>
-            <Route path="contacts" element={<h1>Contacts</h1>} />
-            <Route path="contacts/:id">
-              <Route
-                index
-                element={<UseRouterNavigateWrapper to=".." relative="path" />}
-              />
-            </Route>
-          </>
-        ),
-        { initialEntries: ["/contacts/1"] }
-      );
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(<RouterProvider router={router} />);
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          Contacts
-        </h1>
-      `);
-    });
-
-    it("handles upward navigation from inside a pathless layout route", () => {
-      let router = createMemoryRouter(
-        createRoutesFromElements(
-          <>
-            <Route path="contacts" element={<h1>Contacts</h1>} />
-            <Route element={<Outlet />}>
-              <Route
-                path="contacts/:id"
-                element={<UseRouterNavigateWrapper to=".." relative="path" />}
-              />
-            </Route>
-          </>
-        ),
-        { initialEntries: ["/contacts/1"] }
-      );
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(<RouterProvider router={router} />);
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          Contacts
-        </h1>
-      `);
-    });
-
-    it("handles upward navigation from inside multiple pathless layout routes + index route", () => {
-      let router = createMemoryRouter(
-        createRoutesFromElements(
-          <>
-            <Route path="contacts" element={<h1>Contacts</h1>} />
-            <Route path="contacts/:id">
-              <Route element={<Outlet />}>
-                <Route element={<Outlet />}>
-                  <Route element={<Outlet />}>
-                    <Route
-                      index
-                      element={
-                        <UseRouterNavigateWrapper to=".." relative="path" />
-                      }
-                    />
-                  </Route>
-                </Route>
-              </Route>
-            </Route>
-          </>
-        ),
-        { initialEntries: ["/contacts/1"] }
-      );
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(<RouterProvider router={router} />);
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          Contacts
-        </h1>
-      `);
-    });
-
-    it("handles upward navigation from inside multiple pathless layout routes + path route", () => {
-      let router = createMemoryRouter(
-        createRoutesFromElements(
-          <>
-            <Route path="contacts" element={<Outlet />}>
-              <Route index element={<h1>Contacts</h1>} />
-              <Route element={<Outlet />}>
-                <Route element={<Outlet />}>
-                  <Route element={<Outlet />}>
-                    <Route
-                      path=":id"
-                      element={
-                        <UseRouterNavigateWrapper to=".." relative="path" />
-                      }
-                    />
-                  </Route>
-                </Route>
-              </Route>
-            </Route>
-          </>
-        ),
-        { initialEntries: ["/contacts/1"] }
-      );
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(<RouterProvider router={router} />);
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          Contacts
-        </h1>
-      `);
-    });
-
-    it("handles relative navigation from nested index route", () => {
-      let router = createMemoryRouter(
-        createRoutesFromElements(
-          <>
-            <Route path="layout">
-              <Route path=":param">
-                {/* redirect /layout/:param/ index routes to /layout/:param/dest */}
-                <Route
-                  index
-                  element={
-                    <UseRouterNavigateWrapper to="dest" relative="path" />
-                  }
-                />
-                <Route path="dest" element={<h1>Destination</h1>} />
-              </Route>
-            </Route>
-          </>
-        ),
-        { initialEntries: ["/layout/thing"] }
-      );
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(<RouterProvider router={router} />);
-      });
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        <h1>
-          Destination
-        </h1>
-      `);
-    });
-
-    it("preserves search params and hash", () => {
-      let router = createMemoryRouter(
-        createRoutesFromElements(
-          <>
-            <Route path="contacts" element={<Contacts />} />
-            <Route
-              path="contacts/:id"
-              element={
-                <UseRouterNavigateWrapper
-                  to="..?foo=bar#hash"
-                  relative="path"
-                />
-              }
-            />
-          </>
-        ),
-        { initialEntries: ["/contacts/1"] }
-      );
-      let renderer: TestRenderer.ReactTestRenderer;
-      TestRenderer.act(() => {
-        renderer = TestRenderer.create(<RouterProvider router={router} />);
-      });
-
-      function Contacts() {
-        let { search, hash } = useLocation();
-        return (
-          <>
-            <h1>Contacts</h1>
-            <p>
-              {search}
-              {hash}
-            </p>
-          </>
-        );
-      }
-
-      expect(renderer.toJSON()).toMatchInlineSnapshot(`
-        [
-          <h1>
-            Contacts
-          </h1>,
-          <p>
-            ?foo=bar
-            #hash
-          </p>,
-        ]
-      `);
-    });
-  });
-
-  it("is stable across location changes", () => {
-    let router = createMemoryRouter(
-      [
-        {
-          path: "/",
-          Component: () => (
-            <>
-              <NavBar />
-              <Outlet />
-            </>
-          ),
-          children: [
-            {
-              path: "home",
-              element: <h1>Home</h1>,
-            },
-            {
-              path: "about",
-              element: <h1>About</h1>,
-            },
-          ],
-        },
-      ],
-      { initialEntries: ["/home"] }
-    );
-    let renderer: TestRenderer.ReactTestRenderer;
-    TestRenderer.act(() => {
-      renderer = TestRenderer.create(<RouterProvider router={router} />);
-    });
-
-    function NavBar() {
-      let count = React.useRef(0);
-      let navigate = useRouterNavigate();
-      React.useEffect(() => {
-        count.current++;
-      }, [navigate]);
-      return <p>{count.current}</p>;
-    }
-
-    expect(renderer.toJSON()).toMatchInlineSnapshot(`
-      [
-        <p>
-          0
-        </p>,
-        <h1>
-          Home
-        </h1>,
-      ]
-    `);
-
-    TestRenderer.act(() => {
-      router.navigate("/about");
-    });
-
-    expect(renderer.toJSON()).toMatchInlineSnapshot(`
-      [
-        <p>
-          1
-        </p>,
-        <h1>
-          About
-        </h1>,
-      ]
-    `);
-
-    TestRenderer.act(() => {
-      router.navigate("/home");
-    });
-
-    expect(renderer.toJSON()).toMatchInlineSnapshot(`
-      [
-        <p>
-          1
-        </p>,
-        <h1>
-          Home
-        </h1>,
-      ]
     `);
   });
 });
