+++ packages/react-router/__tests__/useNavigate-test.tsx
@@ -1,11 +1,17 @@
 import * as React from "react";
 import * as TestRenderer from "react-test-renderer";
+import type { RelativeRoutingType, To } from "react-router";
 import {
   MemoryRouter,
   Routes,
   Route,
   useNavigate,
   useLocation,
+  createMemoryRouter,
+  createRoutesFromElements,
+  Outlet,
+  RouterProvider,
+  useRouterNavigate,
 } from "react-router";
 
 describe("useNavigate", () => {
@@ -37,12 +43,11 @@ describe("useNavigate", () => {
       );
     });
 
+    // @ts-expect-error
     let button = renderer.root.findByType("button");
+    TestRenderer.act(() => button.props.onClick());
 
-    TestRenderer.act(() => {
-      button.props.onClick();
-    });
-
+    // @ts-expect-error
     expect(renderer.toJSON()).toMatchInlineSnapshot(`
       <h1>
         About
@@ -74,6 +79,7 @@ describe("useNavigate", () => {
       );
     });
 
+    // @ts-expect-error
     expect(renderer.toJSON()).toMatchInlineSnapshot(`
       [
         <p>
@@ -87,12 +93,11 @@ describe("useNavigate", () => {
       ]
     `);
 
+    // @ts-expect-error
     let button = renderer.root.findByType("button");
+    TestRenderer.act(() => button.props.onClick());
 
-    TestRenderer.act(() => {
-      button.props.onClick();
-    });
-
+    // @ts-expect-error
     expect(renderer.toJSON()).toMatchInlineSnapshot(`
       [
         <p>
@@ -131,6 +136,7 @@ describe("useNavigate", () => {
       );
     });
 
+    // @ts-expect-error
     expect(renderer.toJSON()).toMatchInlineSnapshot(`
       [
         <p>
@@ -144,12 +150,11 @@ describe("useNavigate", () => {
       ]
     `);
 
+    // @ts-expect-error
     let button = renderer.root.findByType("button");
+    TestRenderer.act(() => button.props.onClick());
 
-    TestRenderer.act(() => {
-      button.props.onClick();
-    });
-
+    // @ts-expect-error
     expect(renderer.toJSON()).toMatchInlineSnapshot(`
       [
         <p>
@@ -188,6 +193,7 @@ describe("useNavigate", () => {
       );
     });
 
+    // @ts-expect-error
     expect(renderer.toJSON()).toMatchInlineSnapshot(`
       [
         <p>
@@ -201,12 +207,11 @@ describe("useNavigate", () => {
       ]
     `);
 
+    // @ts-expect-error
     let button = renderer.root.findByType("button");
+    TestRenderer.act(() => button.props.onClick());
 
-    TestRenderer.act(() => {
-      button.props.onClick();
-    });
-
+    // @ts-expect-error
     expect(renderer.toJSON()).toMatchInlineSnapshot(`
       [
         <p>
@@ -330,12 +335,11 @@ describe("useNavigate", () => {
         );
       });
 
+      // @ts-expect-error
       let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
 
-      TestRenderer.act(() => {
-        button.props.onClick();
-      });
-
+      // @ts-expect-error
       expect(renderer.toJSON()).toMatchInlineSnapshot(`
         <p>
           location.state:
@@ -343,5 +347,1276 @@ describe("useNavigate", () => {
         </p>
       `);
     });
+  });
+});
+
+function UseNavigateButton({
+  to,
+  relative,
+}: {
+  to: To;
+  relative?: RelativeRoutingType;
+}) {
+  let navigate = useNavigate();
+  return <button onClick={() => navigate(to, { relative })}>Navigate</button>;
+}
+
+describe("useNavigate (mimicing <Navigate> tests)", () => {
+  describe("with an absolute href", () => {
+    it("navigates to the correct URL", () => {
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(
+          <MemoryRouter initialEntries={["/home"]}>
+            <Routes>
+              <Route path="home" element={<UseNavigateButton to="/about" />} />
+              <Route path="about" element={<h1>About</h1>} />
+            </Routes>
+          </MemoryRouter>
+        );
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          About
+        </h1>
+      `);
+    });
+  });
+
+  describe("with a relative href (relative=route)", () => {
+    it("navigates to the correct URL", () => {
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(
+          <MemoryRouter initialEntries={["/home"]}>
+            <Routes>
+              <Route
+                path="home"
+                element={<UseNavigateButton to="../about" />}
+              />
+              <Route path="about" element={<h1>About</h1>} />
+            </Routes>
+          </MemoryRouter>
+        );
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          About
+        </h1>
+      `);
+    });
+
+    it("handles upward navigation from an index routes", () => {
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(
+          <MemoryRouter initialEntries={["/home"]}>
+            <Routes>
+              <Route path="home">
+                <Route index element={<UseNavigateButton to="../about" />} />
+              </Route>
+              <Route path="about" element={<h1>About</h1>} />
+            </Routes>
+          </MemoryRouter>
+        );
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          About
+        </h1>
+      `);
+    });
+
+    it("handles upward navigation from inside a pathless layout route", () => {
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(
+          <MemoryRouter initialEntries={["/home"]}>
+            <Routes>
+              <Route element={<Outlet />}>
+                <Route
+                  path="home"
+                  element={<UseNavigateButton to="../about" />}
+                />
+              </Route>
+              <Route path="about" element={<h1>About</h1>} />
+            </Routes>
+          </MemoryRouter>
+        );
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          About
+        </h1>
+      `);
+    });
+
+    it("handles upward navigation from inside multiple pathless layout routes + index route", () => {
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(
+          <MemoryRouter initialEntries={["/home"]}>
+            <Routes>
+              <Route path="home">
+                <Route element={<Outlet />}>
+                  <Route element={<Outlet />}>
+                    <Route element={<Outlet />}>
+                      <Route
+                        index
+                        element={<UseNavigateButton to="../about" />}
+                      />
+                    </Route>
+                  </Route>
+                </Route>
+              </Route>
+              <Route path="about" element={<h1>About</h1>} />
+            </Routes>
+          </MemoryRouter>
+        );
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          About
+        </h1>
+      `);
+    });
+
+    it("handles upward navigation from inside multiple pathless layout routes + path route", () => {
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(
+          <MemoryRouter initialEntries={["/home/page"]}>
+            <Routes>
+              <Route path="home" element={<Outlet />}>
+                <Route element={<Outlet />}>
+                  <Route element={<Outlet />}>
+                    <Route element={<Outlet />}>
+                      <Route
+                        path="page"
+                        element={<UseNavigateButton to="../../about" />}
+                      />
+                    </Route>
+                  </Route>
+                </Route>
+              </Route>
+              <Route path="about" element={<h1>About</h1>} />
+            </Routes>
+          </MemoryRouter>
+        );
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          About
+        </h1>
+      `);
+    });
+
+    it("handles parent navigation from inside multiple pathless layout routes", () => {
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(
+          <MemoryRouter initialEntries={["/home/page"]}>
+            <Routes>
+              <Route
+                path="home"
+                element={
+                  <>
+                    <h1>Home</h1>
+                    <Outlet />
+                  </>
+                }
+              >
+                <Route element={<Outlet />}>
+                  <Route element={<Outlet />}>
+                    <Route element={<Outlet />}>
+                      <Route
+                        path="page"
+                        element={
+                          <>
+                            <h2>Page</h2>
+                            <UseNavigateButton to=".." />
+                          </>
+                        }
+                      />
+                    </Route>
+                  </Route>
+                </Route>
+              </Route>
+              <Route path="about" element={<h1>About</h1>} />
+            </Routes>
+          </MemoryRouter>
+        );
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          Home
+        </h1>
+      `);
+    });
+
+    it("handles relative navigation from nested index route", () => {
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(
+          <MemoryRouter initialEntries={["/layout/thing"]}>
+            <Routes>
+              <Route path="layout">
+                <Route path=":param">
+                  {/* redirect /layout/:param/ index routes to /layout/:param/dest */}
+                  <Route index element={<UseNavigateButton to="dest" />} />
+                  <Route path="dest" element={<h1>Destination</h1>} />
+                </Route>
+              </Route>
+            </Routes>
+          </MemoryRouter>
+        );
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          Destination
+        </h1>
+      `);
+    });
+  });
+
+  describe("with a relative href (relative=path)", () => {
+    it("navigates to the correct URL", () => {
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(
+          <MemoryRouter initialEntries={["/contacts/1"]}>
+            <Routes>
+              <Route path="contacts" element={<h1>Contacts</h1>} />
+              <Route
+                path="contacts/:id"
+                element={<UseNavigateButton to=".." relative="path" />}
+              />
+            </Routes>
+          </MemoryRouter>
+        );
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          Contacts
+        </h1>
+      `);
+    });
+
+    it("handles upward navigation from an index routes", () => {
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(
+          <MemoryRouter initialEntries={["/contacts/1"]}>
+            <Routes>
+              <Route path="contacts" element={<h1>Contacts</h1>} />
+              <Route path="contacts/:id">
+                <Route
+                  index
+                  element={<UseNavigateButton to=".." relative="path" />}
+                />
+              </Route>
+            </Routes>
+          </MemoryRouter>
+        );
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          Contacts
+        </h1>
+      `);
+    });
+
+    it("handles upward navigation from inside a pathless layout route", () => {
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(
+          <MemoryRouter initialEntries={["/contacts/1"]}>
+            <Routes>
+              <Route path="contacts" element={<h1>Contacts</h1>} />
+              <Route element={<Outlet />}>
+                <Route
+                  path="contacts/:id"
+                  element={<UseNavigateButton to=".." relative="path" />}
+                />
+              </Route>
+            </Routes>
+          </MemoryRouter>
+        );
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          Contacts
+        </h1>
+      `);
+    });
+
+    it("handles upward navigation from inside multiple pathless layout routes + index route", () => {
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(
+          <MemoryRouter initialEntries={["/contacts/1"]}>
+            <Routes>
+              <Route path="contacts" element={<h1>Contacts</h1>} />
+              <Route path="contacts/:id">
+                <Route element={<Outlet />}>
+                  <Route element={<Outlet />}>
+                    <Route element={<Outlet />}>
+                      <Route
+                        index
+                        element={<UseNavigateButton to=".." relative="path" />}
+                      />
+                    </Route>
+                  </Route>
+                </Route>
+              </Route>
+            </Routes>
+          </MemoryRouter>
+        );
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          Contacts
+        </h1>
+      `);
+    });
+
+    it("handles upward navigation from inside multiple pathless layout routes + path route", () => {
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(
+          <MemoryRouter initialEntries={["/contacts/1"]}>
+            <Routes>
+              <Route path="contacts" element={<Outlet />}>
+                <Route index element={<h1>Contacts</h1>} />
+                <Route element={<Outlet />}>
+                  <Route element={<Outlet />}>
+                    <Route element={<Outlet />}>
+                      <Route
+                        path=":id"
+                        element={<UseNavigateButton to=".." relative="path" />}
+                      />
+                    </Route>
+                  </Route>
+                </Route>
+              </Route>
+            </Routes>
+          </MemoryRouter>
+        );
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          Contacts
+        </h1>
+      `);
+    });
+
+    it("handles relative navigation from nested index route", () => {
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(
+          <MemoryRouter initialEntries={["/layout/thing"]}>
+            <Routes>
+              <Route path="layout">
+                <Route path=":param">
+                  {/* redirect /layout/:param/ index routes to /layout/:param/dest */}
+                  <Route
+                    index
+                    element={<UseNavigateButton to="dest" relative="path" />}
+                  />
+                  <Route path="dest" element={<h1>Destination</h1>} />
+                </Route>
+              </Route>
+            </Routes>
+          </MemoryRouter>
+        );
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          Destination
+        </h1>
+      `);
+    });
+
+    it("preserves search params and hash", () => {
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(
+          <MemoryRouter initialEntries={["/contacts/1"]}>
+            <Routes>
+              <Route path="contacts" element={<Contacts />} />
+              <Route
+                path="contacts/:id"
+                element={
+                  <UseNavigateButton to="..?foo=bar#hash" relative="path" />
+                }
+              />
+            </Routes>
+          </MemoryRouter>
+        );
+      });
+
+      function Contacts() {
+        let { search, hash } = useLocation();
+        return (
+          <>
+            <h1>Contacts</h1>
+            <p>
+              {search}
+              {hash}
+            </p>
+          </>
+        );
+      }
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        [
+          <h1>
+            Contacts
+          </h1>,
+          <p>
+            ?foo=bar
+            #hash
+          </p>,
+        ]
+      `);
+    });
+  });
+
+  it("is not stable across location changes", () => {
+    let router = createMemoryRouter(
+      [
+        {
+          path: "/",
+          Component: () => (
+            <>
+              <NavBar />
+              <Outlet />
+            </>
+          ),
+          children: [
+            {
+              path: "home",
+              element: <h1>Home</h1>,
+            },
+            {
+              path: "about",
+              element: <h1>About</h1>,
+            },
+          ],
+        },
+      ],
+      { initialEntries: ["/home"] }
+    );
+
+    let renderer: TestRenderer.ReactTestRenderer;
+    TestRenderer.act(() => {
+      renderer = TestRenderer.create(<RouterProvider router={router} />);
+    });
+
+    function NavBar() {
+      let count = React.useRef(0);
+      let navigate = useNavigate();
+      React.useEffect(() => {
+        count.current++;
+      }, [navigate]);
+      return <p>{count.current}</p>;
+    }
+
+    // @ts-expect-error
+    expect(renderer.toJSON()).toMatchInlineSnapshot(`
+      [
+        <p>
+          0
+        </p>,
+        <h1>
+          Home
+        </h1>,
+      ]
+    `);
+
+    TestRenderer.act(() => {
+      router.navigate("/about");
+    });
+
+    // @ts-expect-error
+    expect(renderer.toJSON()).toMatchInlineSnapshot(`
+      [
+        <p>
+          1
+        </p>,
+        <h1>
+          About
+        </h1>,
+      ]
+    `);
+
+    TestRenderer.act(() => {
+      router.navigate("/home");
+    });
+
+    // @ts-expect-error
+    expect(renderer.toJSON()).toMatchInlineSnapshot(`
+      [
+        <p>
+          2
+        </p>,
+        <h1>
+          Home
+        </h1>,
+      ]
+    `);
+  });
+});
+
+function UseRouterNavigateButton({
+  to,
+  relative,
+}: {
+  to: To;
+  relative?: RelativeRoutingType;
+}) {
+  let navigate = useRouterNavigate();
+  return <button onClick={() => navigate(to, { relative })}>Navigate</button>;
+}
+
+describe("useRouterNavigate (mimicing <Navigate> tests)", () => {
+  describe("with an absolute href", () => {
+    it("navigates to the correct URL", () => {
+      let router = createMemoryRouter(
+        createRoutesFromElements(
+          <>
+            <Route
+              path="home"
+              element={<UseRouterNavigateButton to="/about" />}
+            />
+            <Route path="about" element={<h1>About</h1>} />
+          </>
+        ),
+        { initialEntries: ["/home"] }
+      );
+
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(<RouterProvider router={router} />);
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          About
+        </h1>
+      `);
+    });
+  });
+
+  describe("with a relative href (relative=route)", () => {
+    it("navigates to the correct URL", () => {
+      let router = createMemoryRouter(
+        createRoutesFromElements(
+          <>
+            <Route
+              path="home"
+              element={<UseRouterNavigateButton to="../about" />}
+            />
+            <Route path="about" element={<h1>About</h1>} />
+          </>
+        ),
+        { initialEntries: ["/home"] }
+      );
+
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(<RouterProvider router={router} />);
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          About
+        </h1>
+      `);
+    });
+
+    it("handles upward navigation from an index routes", () => {
+      let router = createMemoryRouter(
+        createRoutesFromElements(
+          <>
+            <Route path="home">
+              <Route
+                index
+                element={<UseRouterNavigateButton to="../about" />}
+              />
+            </Route>
+            <Route path="about" element={<h1>About</h1>} />
+          </>
+        ),
+        { initialEntries: ["/home"] }
+      );
+
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(<RouterProvider router={router} />);
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          About
+        </h1>
+      `);
+    });
+
+    it("handles upward navigation from inside a pathless layout route", () => {
+      let router = createMemoryRouter(
+        createRoutesFromElements(
+          <>
+            <Route element={<Outlet />}>
+              <Route
+                path="home"
+                element={<UseRouterNavigateButton to="../about" />}
+              />
+            </Route>
+            <Route path="about" element={<h1>About</h1>} />
+          </>
+        ),
+        { initialEntries: ["/home"] }
+      );
+
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(<RouterProvider router={router} />);
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          About
+        </h1>
+      `);
+    });
+
+    it("handles upward navigation from inside multiple pathless layout routes + index route", () => {
+      let router = createMemoryRouter(
+        createRoutesFromElements(
+          <>
+            <Route path="home">
+              <Route element={<Outlet />}>
+                <Route element={<Outlet />}>
+                  <Route element={<Outlet />}>
+                    <Route
+                      index
+                      element={<UseRouterNavigateButton to="../about" />}
+                    />
+                  </Route>
+                </Route>
+              </Route>
+            </Route>
+            <Route path="about" element={<h1>About</h1>} />
+          </>
+        ),
+        { initialEntries: ["/home"] }
+      );
+
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(<RouterProvider router={router} />);
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          About
+        </h1>
+      `);
+    });
+
+    it("handles upward navigation from inside multiple pathless layout routes + path route", () => {
+      let router = createMemoryRouter(
+        createRoutesFromElements(
+          <>
+            <Route path="home" element={<Outlet />}>
+              <Route element={<Outlet />}>
+                <Route element={<Outlet />}>
+                  <Route element={<Outlet />}>
+                    <Route
+                      path="page"
+                      element={<UseRouterNavigateButton to="../../about" />}
+                    />
+                  </Route>
+                </Route>
+              </Route>
+            </Route>
+            <Route path="about" element={<h1>About</h1>} />
+          </>
+        ),
+        { initialEntries: ["/home/page"] }
+      );
+
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(<RouterProvider router={router} />);
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          About
+        </h1>
+      `);
+    });
+
+    it("handles parent navigation from inside multiple pathless layout routes", () => {
+      let router = createMemoryRouter(
+        createRoutesFromElements(
+          <>
+            <Route
+              path="home"
+              element={
+                <>
+                  <h1>Home</h1>
+                  <Outlet />
+                </>
+              }
+            >
+              <Route element={<Outlet />}>
+                <Route element={<Outlet />}>
+                  <Route element={<Outlet />}>
+                    <Route
+                      path="page"
+                      element={
+                        <>
+                          <h2>Page</h2>
+                          <UseRouterNavigateButton to=".." />
+                        </>
+                      }
+                    />
+                  </Route>
+                </Route>
+              </Route>
+            </Route>
+            <Route path="about" element={<h1>About</h1>} />
+          </>
+        ),
+        { initialEntries: ["/home/page"] }
+      );
+
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(<RouterProvider router={router} />);
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          Home
+        </h1>
+      `);
+    });
+
+    it("handles relative navigation from nested index route", () => {
+      let router = createMemoryRouter(
+        createRoutesFromElements(
+          <>
+            <Route path="layout">
+              <Route path=":param">
+                {/* redirect /layout/:param/ index routes to /layout/:param/dest */}
+                <Route index element={<UseRouterNavigateButton to="dest" />} />
+                <Route path="dest" element={<h1>Destination</h1>} />
+              </Route>
+            </Route>
+          </>
+        ),
+        { initialEntries: ["/layout/thing"] }
+      );
+
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(<RouterProvider router={router} />);
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          Destination
+        </h1>
+      `);
+    });
+  });
+
+  describe("with a relative href (relative=path)", () => {
+    it("navigates to the correct URL", () => {
+      let router = createMemoryRouter(
+        createRoutesFromElements(
+          <>
+            <Route path="contacts" element={<h1>Contacts</h1>} />
+            <Route
+              path="contacts/:id"
+              element={<UseRouterNavigateButton to=".." relative="path" />}
+            />
+          </>
+        ),
+        { initialEntries: ["/contacts/1"] }
+      );
+
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(<RouterProvider router={router} />);
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          Contacts
+        </h1>
+      `);
+    });
+
+    it("handles upward navigation from an index routes", () => {
+      let router = createMemoryRouter(
+        createRoutesFromElements(
+          <>
+            <Route path="contacts" element={<h1>Contacts</h1>} />
+            <Route path="contacts/:id">
+              <Route
+                index
+                element={<UseRouterNavigateButton to=".." relative="path" />}
+              />
+            </Route>
+          </>
+        ),
+        { initialEntries: ["/contacts/1"] }
+      );
+
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(<RouterProvider router={router} />);
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          Contacts
+        </h1>
+      `);
+    });
+
+    it("handles upward navigation from inside a pathless layout route", () => {
+      let router = createMemoryRouter(
+        createRoutesFromElements(
+          <>
+            <Route path="contacts" element={<h1>Contacts</h1>} />
+            <Route element={<Outlet />}>
+              <Route
+                path="contacts/:id"
+                element={<UseRouterNavigateButton to=".." relative="path" />}
+              />
+            </Route>
+          </>
+        ),
+        { initialEntries: ["/contacts/1"] }
+      );
+
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(<RouterProvider router={router} />);
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          Contacts
+        </h1>
+      `);
+    });
+
+    it("handles upward navigation from inside multiple pathless layout routes + index route", () => {
+      let router = createMemoryRouter(
+        createRoutesFromElements(
+          <>
+            <Route path="contacts" element={<h1>Contacts</h1>} />
+            <Route path="contacts/:id">
+              <Route element={<Outlet />}>
+                <Route element={<Outlet />}>
+                  <Route element={<Outlet />}>
+                    <Route
+                      index
+                      element={
+                        <UseRouterNavigateButton to=".." relative="path" />
+                      }
+                    />
+                  </Route>
+                </Route>
+              </Route>
+            </Route>
+          </>
+        ),
+        { initialEntries: ["/contacts/1"] }
+      );
+
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(<RouterProvider router={router} />);
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          Contacts
+        </h1>
+      `);
+    });
+
+    it("handles upward navigation from inside multiple pathless layout routes + path route", () => {
+      let router = createMemoryRouter(
+        createRoutesFromElements(
+          <>
+            <Route path="contacts" element={<Outlet />}>
+              <Route index element={<h1>Contacts</h1>} />
+              <Route element={<Outlet />}>
+                <Route element={<Outlet />}>
+                  <Route element={<Outlet />}>
+                    <Route
+                      path=":id"
+                      element={
+                        <UseRouterNavigateButton to=".." relative="path" />
+                      }
+                    />
+                  </Route>
+                </Route>
+              </Route>
+            </Route>
+          </>
+        ),
+        { initialEntries: ["/contacts/1"] }
+      );
+
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(<RouterProvider router={router} />);
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          Contacts
+        </h1>
+      `);
+    });
+
+    it("handles relative navigation from nested index route", () => {
+      let router = createMemoryRouter(
+        createRoutesFromElements(
+          <>
+            <Route path="layout">
+              <Route path=":param">
+                {/* redirect /layout/:param/ index routes to /layout/:param/dest */}
+                <Route
+                  index
+                  element={
+                    <UseRouterNavigateButton to="dest" relative="path" />
+                  }
+                />
+                <Route path="dest" element={<h1>Destination</h1>} />
+              </Route>
+            </Route>
+          </>
+        ),
+        { initialEntries: ["/layout/thing"] }
+      );
+
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(<RouterProvider router={router} />);
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        <h1>
+          Destination
+        </h1>
+      `);
+    });
+
+    it("preserves search params and hash", () => {
+      let router = createMemoryRouter(
+        createRoutesFromElements(
+          <>
+            <Route path="contacts" element={<Contacts />} />
+            <Route
+              path="contacts/:id"
+              element={
+                <UseRouterNavigateButton to="..?foo=bar#hash" relative="path" />
+              }
+            />
+          </>
+        ),
+        { initialEntries: ["/contacts/1"] }
+      );
+
+      function Contacts() {
+        let { search, hash } = useLocation();
+        return (
+          <>
+            <h1>Contacts</h1>
+            <p>
+              {search}
+              {hash}
+            </p>
+          </>
+        );
+      }
+
+      let renderer: TestRenderer.ReactTestRenderer;
+      TestRenderer.act(() => {
+        renderer = TestRenderer.create(<RouterProvider router={router} />);
+      });
+
+      // @ts-expect-error
+      let button = renderer.root.findByType("button");
+      TestRenderer.act(() => button.props.onClick());
+
+      // @ts-expect-error
+      expect(renderer.toJSON()).toMatchInlineSnapshot(`
+        [
+          <h1>
+            Contacts
+          </h1>,
+          <p>
+            ?foo=bar
+            #hash
+          </p>,
+        ]
+      `);
+    });
+  });
+
+  it("is stable across location changes", () => {
+    let router = createMemoryRouter(
+      [
+        {
+          path: "/",
+          Component: () => (
+            <>
+              <NavBar />
+              <Outlet />
+            </>
+          ),
+          children: [
+            {
+              path: "home",
+              element: <h1>Home</h1>,
+            },
+            {
+              path: "about",
+              element: <h1>About</h1>,
+            },
+          ],
+        },
+      ],
+      { initialEntries: ["/home"] }
+    );
+
+    function NavBar() {
+      let count = React.useRef(0);
+      let navigate = useRouterNavigate();
+      React.useEffect(() => {
+        count.current++;
+      }, [navigate]);
+      return <p>{count.current}</p>;
+    }
+
+    let renderer: TestRenderer.ReactTestRenderer;
+    TestRenderer.act(() => {
+      renderer = TestRenderer.create(<RouterProvider router={router} />);
+    });
+
+    // @ts-expect-error
+    expect(renderer.toJSON()).toMatchInlineSnapshot(`
+      [
+        <p>
+          0
+        </p>,
+        <h1>
+          Home
+        </h1>,
+      ]
+    `);
+
+    TestRenderer.act(() => {
+      router.navigate("/about");
+    });
+
+    // @ts-expect-error
+    expect(renderer.toJSON()).toMatchInlineSnapshot(`
+      [
+        <p>
+          1
+        </p>,
+        <h1>
+          About
+        </h1>,
+      ]
+    `);
+
+    TestRenderer.act(() => {
+      router.navigate("/home");
+    });
+
+    // @ts-expect-error
+    expect(renderer.toJSON()).toMatchInlineSnapshot(`
+      [
+        <p>
+          1
+        </p>,
+        <h1>
+          Home
+        </h1>,
+      ]
+    `);
   });
 });
